🔨 refactor: refactored code, bugs fixed, exception handling.
+ Refactored code into many files, now client.py is more clear.
+ Global Logger, no longer unnecessary passing around.
+ Authentication url is not always None anymore.
+ Sometimes Lanis passes incorrect dates and then the lib crashes.
+ Exception handling! PageNotFoundError and NotAuthenticatedError
+ httpx.Post and httpx.Get also with exception handling
   and in a separate class.
